requirements.txt
aiogram
SQLAlchemy
python-dotenv
asyncpg
asyncio
openpyxl